requirements.txt
alembic==1.3.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
